Invert orientation intent
The old method implied that the `h` or `v` represented the direction of the grid. In the months of development for ggn, I've rarely gotten it right the first time, my mental association being that the orientation refers to the guides themselves. Since this seems to be natural, let's use it.